changed game context directory
