+++ src/hooks/useGameContext.tsx
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { GameContext } from "../context/GameContextProvider";
+import { GameContext } from "../context/GameContext";
 
 export const useGameContext = () => {
 	return useContext(GameContext);
